requirements.txt
Flask>=1.1.1
gunicorn>=19.9.0
html5lib>=1.0.1
Jinja2>=2.11.1
json5>=0.9.1
jsonpickle>=1.4.1
jsonschema>=3.2.0
pickleshare>=0.7.5
py>=1.8.1
pyflakes>=2.1.1
requests>=2.22.0
requests-oauthlib>=1.3.0
azure-cognitiveservices-speech>=1.22.0
azure-common>=1.1.25
azure-graphrbac>=0.61.1
azure-mgmt-authorization>=0.60.0
azure-mgmt-containerregistry>=2.8.0
azure-mgmt-keyvault>=2.2.0
azure-mgmt-network>=10.2.0
azure-mgmt-resource>=10.1.0
azure-mgmt-storage>=11.1.0